Placate MSVC in areas of base::size usage
FAILED: obj/third_party/mini_chromium/mini_chromium/base/strings/base.stringprintf.obj
ninja -t msvc -e environment.amd64 -- cl.exe /nologo /showIncludes -DNDEBUG -I../../third_party/mini_chromium/mini_chromium /DNOMINMAX /DUNICODE /DWIN32_LEAN_AND_MEAN /D_CRT_SECURE_NO_WARNINGS /D_HAS_EXCEPTIONS=0 /D_UNICODE /FS /W4 /WX /Zi /bigobj /wd4100 /wd4127 /wd4324 /wd4351 /wd4577 /wd4996 /GL /O2 /Ob2 /Oy- /Zc:inline /d2Zi+  /c ../../third_party/mini_chromium/mini_chromium/base/strings/stringprintf.cc /Foobj/third_party/mini_chromium/mini_chromium/base/strings/base.stringprintf.obj /Fd"obj/third_party/mini_chromium/mini_chromium/base/base_cc.pdb"
../../third_party/mini_chromium/mini_chromium/base/strings/stringprintf.cc(48): error C2220: warning treated as error - no 'object' file generated
../../third_party/mini_chromium/mini_chromium/base/strings/stringprintf.cc(106): note: see reference to function template instantiation 'void base::`anonymous-namespace'::StringAppendVT<std::string>(StringType *,const char *,va_list)' being compiled
        with
        [
            StringType=std::string
        ]
../../third_party/mini_chromium/mini_chromium/base/strings/stringprintf.cc(48): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

FAILED: obj/third_party/mini_chromium/mini_chromium/base/base.logging.obj
ninja -t msvc -e environment.amd64 -- cl.exe /nologo /showIncludes -DNDEBUG -I../../third_party/mini_chromium/mini_chromium /DNOMINMAX /DUNICODE /DWIN32_LEAN_AND_MEAN /D_CRT_SECURE_NO_WARNINGS /D_HAS_EXCEPTIONS=0 /D_UNICODE /FS /W4 /WX /Zi /bigobj /wd4100 /wd4127 /wd4324 /wd4351 /wd4577 /wd4996 /GL /O2 /Ob2 /Oy- /Zc:inline /d2Zi+  /c ../../third_party/mini_chromium/mini_chromium/base/logging.cc /Foobj/third_party/mini_chromium/mini_chromium/base/base.logging.obj /Fd"obj/third_party/mini_chromium/mini_chromium/base/base_cc.pdb"
../../third_party/mini_chromium/mini_chromium/base/logging.cc(76): error C2220: warning treated as error - no 'object' file generated
../../third_party/mini_chromium/mini_chromium/base/logging.cc(76): warning C4267: 'argument': conversion from 'size_t' to 'DWORD', possible loss of data
Reviewed-on: https://chromium-review.googlesource.com/c/1396043

+++ base/logging.cc
@@ -72,8 +72,13 @@ std::string SystemErrorCodeToString(unsigned long error_code) {
   wchar_t msgbuf[256];
   DWORD flags = FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS |
                 FORMAT_MESSAGE_MAX_WIDTH_MASK;
-  DWORD len = FormatMessage(
-      flags, nullptr, error_code, 0, msgbuf, base::size(msgbuf), nullptr);
+  DWORD len = FormatMessage(flags,
+                            nullptr,
+                            error_code,
+                            0,
+                            msgbuf,
+                            static_cast<DWORD>(base::size(msgbuf)),
+                            nullptr);
   if (len) {
     // Most system messages end in a period and a space. Remove the space if
     // it’s there, because the following StringPrintf() includes one.
